+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Current (in progress)
 
 - Fix user deletion performances issues [#3388](https://github.com/opendatateam/udata/pull/3388)
+- Add search to discussions [#3384](https://github.com/opendatateam/udata/pull/3384)
+- Separate metrics for opened discussions [#3370](https://github.com/opendatateam/udata/pull/3370)
 
 ## 10.8.0 (2025-07-22)
 
@@ -24,7 +26,6 @@
 - Do not crash on invalid schemas API JSON [#3363](https://github.com/opendatateam/udata/pull/3363/)
 - Fix flaky test on timed recent Atom feeds [#3365](https://github.com/opendatateam/udata/pull/3365)
 - Allow to remove checksum from resource [#3369](https://github.com/opendatateam/udata/pull/3369)
-- Separate metrics for opened discussions [#3370](https://github.com/opendatateam/udata/pull/3370)
 
 ## 10.6.0 (2025-07-08)