+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Current (in progress)
+## 10.8.1 (2025-07-25)
 
 - Fix user deletion performances issues [#3388](https://github.com/opendatateam/udata/pull/3388)
 - Add search to discussions [#3384](https://github.com/opendatateam/udata/pull/3384)